refactor(review): move all duplicated codes to common component
