+++ src/components/Review/PageBase.tsx
@@ -1,23 +1,34 @@
-import Head from 'next/head'
 import React from 'react'
 
-import { Box, Typography } from '@mui/material'
+import Head from 'next/head'
+import Image from 'next/image'
+import { MDXRemoteSerializeResult } from 'next-mdx-remote'
+import useTranslation from 'next-translate/useTranslation'
+import { useRouter } from 'next/router'
+
+import { Box, Rating, Typography } from '@mui/material'
+import { useColorScheme } from '@mui/material/styles'
 
 import Giscus from '@giscus/react'
 
 import { Title } from '@components/Title'
+import { MDXRenderer } from '@components/MDXRenderer'
+import { BlogMetadata } from '@components/BlogMetadata'
+
+import { useFormattedDate } from '@hooks/useFormattedDate'
+
+import { BaseReviewPostMetadata } from '@constants/review'
 
 import { getWebsiteBaseUrl } from '@utils/getWebsiteBaseUrl'
-import { useColorScheme } from '@mui/material/styles'
-import { useRouter } from 'next/router'
-import { MDXRenderer } from '@components/MDXRenderer'
-import { MDXRemoteSerializeResult } from 'next-mdx-remote'
 
 export interface ReviewPageBaseProps {
   cardUrl: string
   title: string
   children?: React.ReactNode
   source: MDXRemoteSerializeResult
+  readingTime: number
+  metadata: BaseReviewPostMetadata
+  coverImageAspectRatio?: number
 }
 
 export function ReviewPageBase({
@@ -25,9 +36,14 @@ export function ReviewPageBase({
   title,
   children,
   source,
+  readingTime,
+  metadata,
+  coverImageAspectRatio = 1,
 }: ReviewPageBaseProps) {
+  const { t } = useTranslation('review')
   const { mode } = useColorScheme()
   const { locale } = useRouter()
+  const formattedDate = useFormattedDate(metadata.createdAt)
 
   return (
     <div>
@@ -38,6 +54,25 @@ export function ReviewPageBase({
         <meta name="twitter:image" content={cardUrl} />
       </Head>
       <Title withoutMargin>{title}</Title>
+      <Box mt={1} mb="1.3125rem">
+        <BlogMetadata
+          tokens={[
+            formattedDate,
+            t('readingTime', { count: Math.ceil(readingTime) }),
+          ]}
+        />
+      </Box>
+      <Box display="flex" mb={2} justifyContent="center">
+        <Image
+          src={metadata.coverImage}
+          width={250}
+          height={250 * coverImageAspectRatio}
+          alt={metadata.title}
+        />
+      </Box>
+      <Box display="flex" mb={2} justifyContent="center">
+        <Rating defaultValue={metadata.rating / 2} precision={0.5} readOnly />
+      </Box>
       {children}
       <Box component="section">
         <Typography
